Refactored the abc module and added more stream classes
